prime_backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
supabase==2.4.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
pyjwt==2.8.1
cryptography==42.0.2
requests==2.31.0
